Split processing and report related stuff into separate modules.
Refactoring of sanitise.py
