+++ ooni/pipeline/task/sanitise.py
@@ -23,12 +23,11 @@ from __future__ import print_function
 import re
 import os
 import sys
-import hashlib
 import tarfile
 
-from yaml import safe_load_all
 from yaml import safe_dump_all, safe_dump
 from ooni.pipeline import settings
+from ooni.report import Report
 
 
 def archive_report(report_path):
@@ -49,218 +48,6 @@ def list_report_files(directory):
         for filename in filenames:
             if filename.endswith(".yamloo"):
                 yield os.path.join(dirpath, filename)
-
-
-class processor(object):
-    @staticmethod
-    def http_template(entry):
-        if 'requests' not in entry or not entry['requests']:
-            return entry
-        for i, _ in enumerate(entry['requests']):
-            try:
-                del entry['requests'][i]['response']['body']
-            except:
-                pass
-        return entry
-
-    @staticmethod
-    def http_requests(entry):
-        entry['headers_diff'] = list(entry['headers_diff'])
-        return entry
-
-    @staticmethod
-    def scapy_template(entry):
-        try:
-            entry['answered_packets'] = []
-        except:
-            pass
-
-        try:
-            entry['sent_packets'] = []
-        except:
-            pass
-        return entry
-
-    @staticmethod
-    def dns_template(entry):
-        return entry
-
-    @staticmethod
-    def dns_consistency(entry):
-        entry['tampering'] = entry['tampering'].items()
-        return entry
-
-    @staticmethod
-    def captive_portal(entry):
-        entry['vendor_dns_tests']['google_dns_cp'] = \
-            list(entry['vendor_dns_tests']['google_dns_cp'])
-        return entry
-
-    @staticmethod
-    def null(entry):
-        return entry
-
-    @staticmethod
-    def bridge_reachability_tcp_connect(entry):
-        if entry['input'].strip() in settings.bridge_db_mapping.keys():
-            b = settings.bridge_db_mapping[entry['input'].strip()]
-            fingerprint = b['fingerprint'].decode('hex')
-            hashed_fingerprint = hashlib.sha1(fingerprint).hexdigest()
-            entry['bridge_hashed_fingerprint'] = hashed_fingerprint
-            entry['input'] = hashed_fingerprint
-            return entry
-        return entry
-
-    @staticmethod
-    def bridge_reachability(entry):
-        if entry.get('bridge_address') and \
-                entry['bridge_address'].strip() in settings.bridge_db_mapping:
-            b = settings.bridge_db_mapping[entry['bridge_address'].strip()]
-            entry['distributor'] = b['distributor']
-            fingerprint = b['fingerprint'].decode('hex')
-            hashed_fingerprint = hashlib.sha1(fingerprint).hexdigest()
-            entry['input'] = hashed_fingerprint
-            entry['bridge_address'] = None
-        else:
-            entry['distributor'] = None
-            hashed_fingerprint = None
-            pass
-
-        entry['bridge_hashed_fingerprint'] = hashed_fingerprint
-
-        return entry
-
-    @staticmethod
-    def tcp_connect(entry):
-        entry = processor.bridge_reachability(entry)
-        return entry
-
-    @staticmethod
-    def default(entry):
-        if 'report' in entry:
-            entry = entry['report']
-        return entry
-
-report_processor = {
-    "http_host": processor.http_template,
-    "HTTP Host": processor.http_template,
-
-    "http_requests_test": [processor.http_template, processor.http_requests],
-    "http_requests": [processor.http_template, processor.http_requests],
-    "HTTP Requests Test": [processor.http_template, processor.http_requests],
-
-    "bridge_reachability": processor.bridge_reachability,
-    "bridgereachability": processor.bridge_reachability,
-
-    "TCP Connect": processor.tcp_connect,
-    "tcp_connect": processor.tcp_connect,
-
-    "DNS tamper": [processor.dns_template, processor.dns_consistency],
-    "dns_consistency": [processor.dns_template, processor.dns_consistency],
-
-    "HTTP Invalid Request Line": processor.null,
-    "http_invalid_request_line": processor.null,
-
-    "http_header_field_manipulation": processor.null,
-    "HTTP Header Field Manipulation": processor.null,
-
-    "Multi Protocol Traceroute Test": [processor.scapy_template],
-    "multi_protocol_traceroute_test": [processor.scapy_template],
-    "traceroute": [processor.scapy_template],
-
-    "parasitic_traceroute_test": processor.null,
-
-    "tls-handshake": processor.null,
-
-    "dns_injection": processor.null,
-
-    "captivep": processor.captive_portal,
-    "captiveportal": processor.captive_portal,
-
-    # These are ignored as we don't yet have analytics for them
-    "HTTPFilteringBypass": False,
-    "HTTPTrix": False,
-    "http_test": False,
-    "http_url_list": False,
-    "dns_spoof": False,
-    "netalyzrwrapper": False,
-
-    # These are ignored because not code for them is available
-    "tor_http_requests_test": False,
-    "sip_requests_test": False,
-    "tor_exit_ip_test": False,
-    "website_probe": False,
-    "base_tcp_test": False,
-
-    # These are ignored because they are invalid reports
-    "summary": False,
-    "test_get": False,
-    "test_put": False,
-    "test_post": False,
-    "this_test_is_nameless": False,
-    "test_send_host_header": False,
-    "test_random_big_request_method": False,
-    "test_get_random_capitalization": False,
-    "test_put_random_capitalization": False,
-    "test_post_random_capitalization": False,
-    "test_random_invalid_field_count": False,
-    "keyword_filtering_detection_based_on_rst_packets": False
-}
-
-
-class Report(object):
-    def __init__(self, path):
-        self.fh = open(path)
-        self._report = safe_load_all(self.fh)
-        self.report_path = path
-        self.header = self._report.next()
-
-    def process(self, entry):
-        entry = processor.default(entry)
-        try:
-            ps = report_processor[self.header['test_name']]
-        except:
-            print("Unknown processor for %s: %s" %
-                  (self.header['test_name'], self.report_path))
-            ps = False
-
-        if isinstance(ps, list):
-            for p in ps:
-                try:
-                    entry = p(entry)
-                except:
-                    entry = entry
-            return entry
-        elif hasattr(ps, '__call__'):
-            try:
-                return ps(entry)
-            except:
-                return entry
-        elif ps is False:
-            return False
-        else:
-            raise Exception("Invalid definition of processor of the test")
-
-    def next_entry(self):
-        try:
-            entry = self._report.next()
-        except StopIteration:
-            raise StopIteration
-        except Exception:
-            self.next_entry()
-        if not entry:
-            entry = self.next_entry()
-        entry = self.process(entry)
-        return entry
-
-    def next(self):
-        return self.next_entry()
-
-    def __iter__(self):
-        return self
-
-    def close(self):
-        self.fh.close()
 
 
 def main():
